+++ fabric-loot-api-v3-bridge/src/main/java/dev/loaderbridge/fabric/api/loot/FabricLootApiV3BridgeMod.java
@@ -1,0 +1,6 @@
+package dev.loaderbridge.fabric.api.loot;
+
+import net.minecraftforge.fml.common.Mod;
+
+@Mod("loaderbridge_fabric_loot_api_v3")
+public final class FabricLootApiV3BridgeMod {}